Move block proof to proto domain
